Move common code from LeanPeak to BasePeak
